+++ odoo/local-src/specific_crm/models/__init__.py
@@ -7,3 +7,4 @@ from . import project_specs
 from . import crm_lead
 from . import survey
 from . import crm_team
+from . import project_task