+++ odoo/songs/install/roctool_vars.py
@@ -7,11 +7,8 @@ coa_dict = {
     'base.main_company': 'l10n_fr.l10n_fr_pcg_chart_template',
     '__setup__.roctool_inc': 'l10n_generic_coa.configurable_chart_template',
     '__setup__.roctool_gmbh': 'l10n_de_skr03.l10n_de_chart_template',
-    # '__setup__.roctool_taiwan': l10n_cn,
-    # '__setup__.roctool_japan': 'l10n_jp.l10n_jp1',
-    # cannot import Japan now because of this bug:
-    # https://github.com/odoo/odoo/issues/15384
-    # OPW reported on 2017-02-07
+    # '__setup__.roctool_taiwan': 'l10n_jp.l10n_jp1',
+    '__setup__.roctool_japan': 'l10n_jp.l10n_jp1',
 }
 
 crm_stages = [